+++ fr-server/src/main/resources/templates/mail/ftl/mail.html
@@ -44,7 +44,7 @@
 														尊敬的用户您好！
 														<br/>
 														<br/>
-														来自 Albania 的 Mr hello 先生在网站后台留下了联系方式！
+														来自 <span id="address"></span> 的 <span id="user"></span> 先生在网站后台留下了联系方式！
 														请尽快联系他！
 													
 													</td>
